+++ src/pages/Home.js
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { getAqi } from '../redux/airquality/airquality';
-import AqiDetail from './AqiDetail';
+import Aqi from './Aqi';
 import AqiInput from '../components/AqiInput';
 import { getLocation } from '../redux/getLocation/getLocation';
 
@@ -8,7 +8,6 @@ const Home = () => {
   const dispatch = useDispatch();
   const aqiData = useSelector((state) => state.AqiSlice.aqi);
   const error = useSelector((state) => state.AqiSlice.error);
-  const apilocation = useSelector((state) => state.LocationSlice);
   const inputLocation = (location) => {
     dispatch(getAqi(location));
     dispatch(getLocation(location));
@@ -16,9 +15,8 @@ const Home = () => {
   return (
     <div>
       <AqiInput inputLocation={inputLocation} />
-      <AqiDetail
+      <Aqi
         aqiData={aqiData}
-        location={apilocation}
         error={error}
       />
     </div>
